+++ src/Modules/Core/DBConstants/Region/RegionIDs.php
@@ -28,6 +28,7 @@ class RegionIDs
 	public const SWITZERLAND = 106;
 	public const SWITZERLAND_BOT_GROUP = 1763;
 	public const VOTING_ADMIN_GROUP = 3871;
+	public const BOARD_ADMIN_GROUP = 3875;
 	public const ORGA_COORDINATION_GROUP = 3818;
 
 	// groups used for displaying team page: